@@ -10,17 +10,17 @@
 <hr />
 <div class="row">
     <div class="col-md-4">
-        <form asp-action="Create" enctype="multipart/form-data">
+        <form asp-action="Create">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <div class="form-group">
-                <label asp-for="ImagePath" class="control-label"></label>
-                <input asp-for="ImagePath" class="image/*" />
-                <span asp-validation-for="ImagePath" class="text-danger"></span>
+                <label asp-for="Title" class="control-label"></label>
+                <input asp-for="Title" class="form-control" />
+                <span asp-validation-for="Title" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <label asp-for="ImageFile" class="control-label"></label>
-                <input asp-for="ImageFile" accept="form-control" />
-                <span asp-validation-for="ImageFile" class="text-danger"></span>
+                <label asp-for="ImagePath" class="control-label"></label>
+                <input asp-for="ImagePath" class="form-control" />
+                <span asp-validation-for="ImagePath" class="text-danger"></span>
             </div>
             <div class="form-group">
                 <label asp-for="MemberId" class="control-label"></label>
@@ -29,11 +29,11 @@
             </div>
             <div class="form-group">
                 <label asp-for="FamilyId" class="control-label"></label>
-                <select asp-for="FamilyId" class="form-control" asp-items="ViewBag.FamilyId"></select>
+                <select asp-for="FamilyId" class ="form-control" asp-items="ViewBag.FamilyId"></select>
             </div>
             <div class="form-group">
                 <label asp-for="AlbumId" class="control-label"></label>
-                <select asp-for="AlbumId" class="form-control" asp-items="ViewBag.AlbumId"></select>
+                <select asp-for="AlbumId" class ="form-control" asp-items="ViewBag.AlbumId"></select>
             </div>
             <div class="form-group">
                 <div class="checkbox">
